requirements.txt
python-telegram-bot[job-queue]==22.7
fastapi==0.115.12
uvicorn[standard]==0.34.0
supabase==2.28.2
pydantic-settings==2.13.1
pytz==2024.1
python-dotenv==1.0.0
requests==2.32.5
httpx==0.28.1
faster-whisper==1.2.1
